separated code into multiple files; trying to provide sync and async variant from the same code and to isolate libeio code; segfault on sync variant
