+++ xapian-binding.cc
@@ -7,256 +7,13 @@
 #include <node_events.h>
 #include <node_buffer.h>
 
+#include "xapian-op.h"
+#include "xapian-enquire.h"
+#include "xapian-database.h"
+
 using namespace v8;
 using namespace node;
 
-static Persistent<String> kBusyMsg;
-
-struct AsyncOpBase {
-  AsyncOpBase(Handle<Function> cb)
-    : callback(), error(NULL) {
-    callback = Persistent<Function>::New(cb);
-    ev_ref(EV_DEFAULT_UC);
-  }
-  virtual ~AsyncOpBase() {
-    if (error) delete error;
-    ev_unref(EV_DEFAULT_UC);
-    callback.Dispose();
-  }
-  Persistent<Function> callback;
-  Xapian::Error* error;
-};
-
-template <class T>
-struct AsyncOp : public AsyncOpBase {
-  AsyncOp(Handle<Object> ob, Handle<Function> cb)
-    : AsyncOpBase(cb), object(ObjectWrap::Unwrap<T>(ob)) {
-    if (object->mBusy)
-      throw Exception::Error(kBusyMsg);
-    object->mBusy = true;
-    object->Ref();
-  }
-  virtual ~AsyncOp() { object->Unref(); }
-  void poolDone() { object->mBusy = false; }
-  T* object;
-};
-
-class Database : public EventEmitter {
-public:
-  static void Init(Handle<Object> target);
-
-  static Persistent<FunctionTemplate> constructor_template;
-
-  Xapian::Database& getDb() { return *mDb; }
-
-protected:
-  Database() : EventEmitter(), mDb(NULL), mBusy(false) {}
-
-  virtual ~Database() {
-    if (mDb) {
-      mDb->close();
-      delete mDb;
-    }
-  }
-
-  union {
-    Xapian::Database* mDb;
-    Xapian::WritableDatabase* mWdb;
-  };
-  bool mBusy;
-
-  friend struct AsyncOp<Database>;
-
-  static Handle<Value> New(const Arguments& args);
-
-  static Handle<Value> AddDatabase(const Arguments& args);
-
-  static Handle<Value> Reopen(const Arguments& args);
-  static int Open_pool(eio_req *req);
-  static int Open_done(eio_req *req);
-  struct Open_data : AsyncOp<Database> {
-    Open_data(Handle<Object> ob, Handle<String> file, int wop=0)
-      : AsyncOp<Database>(ob, Handle<Function>()), filename(file), writeopts(wop) {}
-    String::Utf8Value filename;
-    int writeopts;
-  };
-};
-
-class WritableDatabase : public Database {
-public:
-  static void Init(Handle<Object> target);
-
-  static Persistent<FunctionTemplate> constructor_template;
-
-  Xapian::WritableDatabase& getWdb() { return *mWdb; }
-
-protected:
-  WritableDatabase() : Database() { }
-
-  ~WritableDatabase() { }
-
-  friend struct AsyncOp<WritableDatabase>;
-
-  static Handle<Value> New(const Arguments& args);
-
-  static Handle<Value> ReplaceDocument(const Arguments& args);
-  static int AddDocument_pool(eio_req *req);
-  static int AddDocument_done(eio_req *req);
-  struct AddDocument_data : AsyncOp<WritableDatabase> {
-    AddDocument_data(Handle<Object> ob, Handle<Function> cb, const Xapian::Document& doc, Handle<String> id)
-      : AsyncOp<WritableDatabase>(ob, cb), document(doc), idterm(id) {}
-    const Xapian::Document& document;
-    Xapian::docid docid;
-    String::Utf8Value idterm;
-  };
-
-  static Handle<Value> Commit(const Arguments& args);
-  static Handle<Value> BeginTransaction(const Arguments& args);
-  static Handle<Value> CommitTransaction(const Arguments& args);
-
-  static int Commit_pool(eio_req *req);
-  static int Commit_done(eio_req *req);
-  struct Commit_data : AsyncOp<WritableDatabase> {
-    Commit_data(Handle<Object> ob, Handle<Function> cb, int op, bool fl=false)
-      : AsyncOp<WritableDatabase>(ob, cb), type(op), flush(fl) {}
-    enum { eCommit, eBeginTx, eCommitTx };
-    int type;
-    bool flush;
-  };
-};
-
-class TermGenerator : public ObjectWrap {
-public:
-  static void Init(Handle<Object> target);
-
-  static Persistent<FunctionTemplate> constructor_template;
-
-  Xapian::TermGenerator mTg;
-
-protected:
-  TermGenerator() : ObjectWrap(), mTg() { }
-
-  ~TermGenerator() { }
-
-  friend struct Main_data;
-
-  static Handle<Value> New(const Arguments& args);
-
-  static Handle<Value> SetDatabase(const Arguments& args);
-  static Handle<Value> SetFlags(const Arguments& args);
-  static Handle<Value> SetStemmer(const Arguments& args);
-};
-
-class Stem : public ObjectWrap {
-public:
-  static void Init(Handle<Object> target);
-
-  static Persistent<FunctionTemplate> constructor_template;
-
-  Xapian::Stem mStem;
-
-protected:
-  Stem(const char* iLang) : ObjectWrap(), mStem(iLang) { }
-
-  ~Stem() { }
-
-  static Handle<Value> New(const Arguments& args);
-};
-
-class Enquire : public ObjectWrap {
-public:
-  static void Init(Handle<Object> target);
-
-  static Persistent<FunctionTemplate> constructor_template;
-
-protected:
-  Enquire(const Xapian::Database& iDb) : ObjectWrap(), mEnq(iDb), mBusy(false) {}
-
-  ~Enquire() {
-  }
-
-  Xapian::Enquire mEnq;
-  bool mBusy;
-
-  friend struct AsyncOp<Enquire>;
-
-  static Handle<Value> New(const Arguments& args);
-
-  static Handle<Value> SetQuery(const Arguments& args);
-
-  static Handle<Value> GetMset(const Arguments& args);
-  static int GetMset_pool(eio_req *req);
-  static int GetMset_done(eio_req *req);
-  struct GetMset_data : AsyncOp<Enquire> {
-    GetMset_data(Handle<Object> ob, Handle<Function> cb, uint32_t fi, uint32_t mx)
-      : AsyncOp<Enquire>(ob, cb), first(fi), maxitems(mx), set(NULL) {}
-    ~GetMset_data() { if (set) delete [] set; }
-    Xapian::doccount first, maxitems;
-    struct Item {
-      Xapian::docid id;
-      Xapian::Document* document;
-      Xapian::doccount rank, collapse_count;
-      Xapian::weight weight;
-      std::string collapse_key, description;
-      Xapian::percent percent;
-    };
-    Item* set;
-    int size;
-  };
-};
-
-class Query : public ObjectWrap {
-public:
-  static void Init(Handle<Object> target);
-
-  static Persistent<FunctionTemplate> constructor_template;
-
-  Xapian::Query mQry;
-
-protected:
-  template <class T>
-  Query(Xapian::Query::op o, T a, T b) : ObjectWrap(), mQry(o, a, b) {}
-
-  ~Query() {}
-
-  static Handle<Value> New(const Arguments& args);
-
-  //static Handle<Value> Fn(const Arguments& args);
-};
-
-class Document : public ObjectWrap {
-public:
-  static void Init(Handle<Object> target);
-
-  static Persistent<FunctionTemplate> constructor_template;
-
-  Xapian::Document* getDoc() {
-    return mDoc;
-  }
-
-protected:
-  Document(Xapian::Document* iDoc) : ObjectWrap(), mDoc(iDoc), mBusy(false) {}
-
-  ~Document() {
-    delete mDoc;
-  }
-
-  Xapian::Document* mDoc;
-  bool mBusy;
-
-  friend struct AsyncOp<Document>;
-
-  static Handle<Value> New(const Arguments& args);
-
-  static Handle<Value> GetData(const Arguments& args);
-  static int GetData_pool(eio_req *req);
-  static int GetData_done(eio_req *req);
-  struct GetData_data : AsyncOp<Document> {
-    GetData_data(Handle<Object> ob, Handle<Function> cb)
-      : AsyncOp<Document>(ob, cb) {}
-    std::string data;
-  };
-};
 
 class Mime2Text : public ObjectWrap {
 public:
@@ -331,7 +88,7 @@ void init (Handle<Object> target) {
   Mime2Text::Init(target);
   target->Set(String::NewSymbol("assemble_document"), FunctionTemplate::New(AssembleDocument)->GetFunction());
 }
-
+/*
 static void tryCallCatch(Handle<Function> fn, Handle<Object> context, int argc, Handle<Value>* argv) {
   TryCatch try_catch;
 
@@ -347,6 +104,7 @@ static T* GetInstance(Handle<Value> val) {
     return ObjectWrap::Unwrap<T>(val->ToObject());
   return NULL;
 }
+*/
 
 Persistent<FunctionTemplate> Database::constructor_template;
 
@@ -712,121 +470,7 @@ Handle<Value> Stem::New(const Arguments& args) {
   return args.This();
 }
 
-Persistent<FunctionTemplate> Enquire::constructor_template;
 
-void Enquire::Init(Handle<Object> target) {
-  constructor_template = Persistent<FunctionTemplate>::New(FunctionTemplate::New(New));
-  constructor_template->InstanceTemplate()->SetInternalFieldCount(1);
-  constructor_template->SetClassName(String::NewSymbol("Enquire"));
-
-  NODE_SET_PROTOTYPE_METHOD(constructor_template, "set_query", SetQuery);
-  NODE_SET_PROTOTYPE_METHOD(constructor_template, "get_mset", GetMset);
-
-  target->Set(String::NewSymbol("Enquire"), constructor_template->GetFunction());
-}
-
-Handle<Value> Enquire::New(const Arguments& args) {
-  HandleScope scope;
-  Database* aDb;
-  if (args.Length() < 1 || !(aDb = GetInstance<Database>(args[0])))
-    return ThrowException(Exception::TypeError(String::New("arguments are (Database)")));
-  Enquire* that = new Enquire(aDb->getDb());
-  that->Wrap(args.This());
-  return args.This();
-}
-
-Handle<Value> Enquire::SetQuery(const Arguments& args) {
-  HandleScope scope;
-  Query* aQ;
-  if (args.Length() < 1 || !(aQ = GetInstance<Query>(args[0])))
-    return ThrowException(Exception::TypeError(String::New("arguments are (Query)")));
-  Enquire* that = ObjectWrap::Unwrap<Enquire>(args.This());
-  if (that->mBusy)
-    return ThrowException(Exception::Error(kBusyMsg));
-  try {
-    that->mEnq.set_query(aQ->mQry);
-  } catch (const Xapian::Error& err) {
-    return ThrowException(Exception::Error(String::New(err.get_msg().c_str())));
-  }
-  return Undefined();
-}
-
-Handle<Value> Enquire::GetMset(const Arguments& args) {
-  HandleScope scope;
-
-  if (args.Length() < 3 || !args[0]->IsUint32() || !args[1]->IsUint32() || !args[2]->IsFunction())
-    return ThrowException(Exception::TypeError(String::New("arguments are (number, number, function)")));
-  GetMset_data* aData;
-  try {
-    aData = new GetMset_data(args.This(), Local<Function>::Cast(args[2]), args[0]->Uint32Value(), args[1]->Uint32Value());
-  } catch (Local<Value> ex) {
-    return ThrowException(ex);
-  }
-
-  eio_custom(GetMset_pool, EIO_PRI_DEFAULT, GetMset_done, aData);
-
-  return Undefined();
-}
-
-int Enquire::GetMset_pool(eio_req *req) {
-  GetMset_data* aData = (GetMset_data*) req->data;
-
-  try {
-  Xapian::MSet aSet = aData->object->mEnq.get_mset(aData->first, aData->maxitems);
-  aData->set = new GetMset_data::Item[aSet.size()];
-  aData->size = 0;
-  for (Xapian::MSetIterator a = aSet.begin(); a != aSet.end(); ++a, ++aData->size) {
-    aData->set[aData->size].id = *a;
-    aData->set[aData->size].document = new Xapian::Document(a.get_document());
-    aData->set[aData->size].rank = a.get_rank();
-    aData->set[aData->size].collapse_count = a.get_collapse_count();
-    aData->set[aData->size].weight = a.get_weight();
-    aData->set[aData->size].collapse_key = a.get_collapse_key();
-    aData->set[aData->size].description = a.get_description();
-    aData->set[aData->size].percent = a.get_percent();
-  }
-  } catch (const Xapian::Error& err) {
-    aData->error = new Xapian::Error(err);
-  }
-
-  aData->poolDone();
-  return 0;
-}
-
-int Enquire::GetMset_done(eio_req *req) {
-  HandleScope scope;
-
-  GetMset_data* aData = (GetMset_data*) req->data;
-
-  Handle<Value> argv[2];
-  if (aData->error) {
-    argv[0] = Exception::Error(String::New(aData->error->get_msg().c_str()));
-  } else {
-    argv[0] = Null();
-    Local<Array> aList(Array::New(aData->size));
-    Local<Function> aCtor(Document::constructor_template->GetFunction());
-    for (int a = 0; a < aData->size; ++a) {
-      Local<Object> aO(Object::New());
-      Local<Value> aDoc[] = { External::New(aData->set[a].document) };
-      aO->Set(String::NewSymbol("document"      ), aCtor->NewInstance(1, aDoc));
-      aO->Set(String::NewSymbol("id"            ), Uint32::New(aData->set[a].id                  ));
-      aO->Set(String::NewSymbol("rank"          ), Uint32::New(aData->set[a].rank                ));
-      aO->Set(String::NewSymbol("collapse_count"), Uint32::New(aData->set[a].collapse_count      ));
-      aO->Set(String::NewSymbol("weight"        ), Number::New(aData->set[a].weight              ));
-      aO->Set(String::NewSymbol("collapse_key"  ), String::New(aData->set[a].collapse_key.c_str()));
-      aO->Set(String::NewSymbol("description"   ), String::New(aData->set[a].description.c_str() ));
-      aO->Set(String::NewSymbol("percent"       ),  Int32::New(aData->set[a].percent             ));
-      aList->Set(a, aO);
-    }
-    argv[1] = aList;
-  }
-
-  tryCallCatch(aData->callback, aData->object->handle_, aData->error ? 1 : 2, argv);
-
-  delete aData;
-
-  return 0;
-}
 
 Persistent<FunctionTemplate> Query::constructor_template;
