+++ xapian-op.h
@@ -1,0 +1,75 @@
+#ifndef _XAPIAN_OP_
+#define _XAPIAN_OP_
+
+#include <xapian.h>
+#include "mime2text.h"
+
+#include <v8.h>
+#include <node.h>
+#include <node_events.h>
+#include <node_buffer.h>
+
+using namespace v8;
+using namespace node;
+
+static void tryCallCatch(Handle<Function> fn, Handle<Object> context, int argc, Handle<Value>* argv) {
+  TryCatch try_catch;
+
+  fn->Call(context, argc, argv);
+
+  if (try_catch.HasCaught())
+    FatalException(try_catch);
+}
+
+template <class T>
+static T* GetInstance(Handle<Value> val) {
+  if (val->IsObject() && T::constructor_template->HasInstance(val->ToObject()))
+    return ObjectWrap::Unwrap<T>(val->ToObject());
+  return NULL;
+}
+
+static void sendToThreadPool(void* execute, void* done, void* data){
+  eio_custom((eio_cb) execute, EIO_PRI_DEFAULT, (eio_cb) done, data);
+}
+
+static Persistent<String> kBusyMsg;
+
+struct AsyncOpBase {
+  AsyncOpBase(Handle<Function> cb)
+    : callback(), error(NULL) {
+    callback = Persistent<Function>::New(cb);
+    ev_ref(EV_DEFAULT_UC);
+  }
+  virtual ~AsyncOpBase() {
+    if (error) delete error;
+    ev_unref(EV_DEFAULT_UC);
+    callback.Dispose();
+  }
+  Persistent<Function> callback;
+  Xapian::Error* error;
+};
+
+template <class T>
+struct AsyncOp : public AsyncOpBase {
+  AsyncOp(Handle<Object> ob, Handle<Function> cb)
+    : AsyncOpBase(cb), object(ObjectWrap::Unwrap<T>(ob)) {
+    if (object->mBusy)
+      throw Exception::Error(kBusyMsg);
+    object->mBusy = true;
+    object->Ref();
+  }
+  virtual ~AsyncOp() { object->Unref(); }
+  void poolDone() { object->mBusy = false; }
+  T* object;
+};
+
+struct OpInfo {
+  OpInfo(void* data): data(data), op(NULL){}
+  OpInfo(void* data, void *asop): data(data), op(asop) {}
+  void* data;
+  void* op;
+};
+
+
+
+#endif
